Move test syntax classes to their proper packages
