+++ src/test/java/org/skriptlang/skript/test/registration/EvtTestCase.java
@@ -16,7 +16,7 @@
  *
  * Copyright Peter Güttinger, SkriptLang team and contributors
  */
-package ch.njol.skript.test.runner;
+package org.skriptlang.skript.test.registration;
 
 import org.bukkit.Location;
 import org.bukkit.block.Block;
@@ -30,6 +30,10 @@ import ch.njol.skript.lang.Literal;
 import ch.njol.skript.lang.SkriptEvent;
 import ch.njol.skript.lang.SkriptParser;
 import ch.njol.skript.registrations.EventValues;
+import ch.njol.skript.test.runner.SkriptJUnitTest;
+import ch.njol.skript.test.runner.SkriptTestEvent;
+import ch.njol.skript.test.runner.TestMode;
+import ch.njol.skript.test.runner.TestTracker;
 import ch.njol.skript.util.Getter;
 
 public class EvtTestCase extends SkriptEvent {
